+++ Box.V2/Models/BoxFolderLockOperations.cs
@@ -1,0 +1,26 @@
+using System;
+using Newtonsoft.Json;
+
+namespace Box.V2.Models
+{
+    /// <summary>
+    /// Box representation of a operations that have been locked on a folder lock in box
+    /// </summary>
+    public class BoxFolderLockOperations
+    {
+        public const string FieldDelete = "delete";
+        public const string FieldMove = "move";
+
+        /// <summary>
+        /// Whether deleting the folder is restricted
+        /// </summary>
+        [JsonProperty(PropertyName = FieldDelete)]
+        public bool Delete { get; private set; }
+
+        /// <summary>
+        /// Whether deleting the folder is restricted
+        /// </summary>
+        [JsonProperty(PropertyName = FieldMove)]
+        public bool Move { get; private set; }
+    }
+}